+++ packages/react-jss/src/index.test.tsx
@@ -1,0 +1,24 @@
+import React, {Component} from 'react'
+
+import WithStyles from '.'
+
+interface Props {
+  testProp: string
+}
+
+class TestComponent extends Component<Props> {
+  static defaultProps: Props = {
+    testProp: 'hello'
+  }
+
+  render() {
+    return <div>{this.props.testProp}</div>
+  }
+}
+
+const TestComponentWitStyles = WithStyles({})(TestComponent)
+
+function testRender() {
+  // component shouldn't ask to pass `testProp`
+  return <TestComponentWitStyles />
+}